moved types to one file under modules, moved actions back to actions folders and cleaned them up
