+++ packages/augur-ui/src/modules/auth/actions/load-account-data.ts
@@ -1,5 +1,5 @@
 import { loadAccountDataFromLocalStorage } from "modules/auth/actions/load-account-data-from-local-storage";
-import { updateLoginAccountAction } from "modules/common/types/login-account";
+import { updateLoginAccountAction } from "modules/account/actions/login-account";
 import { checkAccountAllowance } from "modules/auth/actions/approve-account";
 import { loadAccountHistory } from "modules/auth/actions/load-account-history";
 import { updateAssets } from "modules/auth/actions/update-assets";
