+++ packages/augur-ui/src/modules/auth/reducers/login-account.ts
@@ -1,11 +1,12 @@
 import { RESET_STATE } from "modules/app/actions/reset-state";
-import { LOGIN_ACTIONS, LoginAccount, LoginAccountAction } from "modules/common/types/login-account";
+import { LOGIN_ACTIONS } from "modules/account/actions/login-account";
+import { LoginAccount, BaseAction } from "modules/types";
 
 const DEFAULT_STATE: LoginAccount = { eth: undefined, rep: undefined, dai: undefined };
 
 export default function(
   loginAccount: LoginAccount = DEFAULT_STATE,
-  action: LoginAccountAction,
+  action: BaseAction,
 ) {
   switch (action.type) {
     case LOGIN_ACTIONS.UPDATE_LOGIN_ACCOUNT:
