--- packages/augur-ui/src/modules/common/types/index.ts
@@ -1,8 +1,0 @@
-// import { ThunkAction, ThunkDispatch } from "redux-thunk";
-import { Action } from "redux";
-
-export type voidFunction = () => void;
-
-export class BaseAction implements Action {
-  public type!: string;
-}
