+++ packages/augur-ui/src/modules/types.ts
@@ -1,0 +1,51 @@
+
+export interface AppStatus {
+  isLogged: boolean|undefined;
+  edgeLoading: boolean|undefined;
+  edgeContext: string|undefined;
+  isConnectionTrayOpen: boolean|undefined;
+}
+
+export interface PositionData {
+  marketId: string;
+  netPosition: string;
+  outcome: string;
+  position: string;
+  averagePrice: string;
+  realized: string;
+  timestamp: number;
+  total: string;
+  unrealized: string;
+  cost: string;
+  unrealizedCost: string;
+  unrealizedRevenue: string;
+  totalPercent: string;
+  unrealizedPercent: string;
+  realizedPercent: string;
+  unrealizedRevenue24hChangePercent: string;
+}
+
+export interface AccountPosition {
+  marketId: string | null;
+  positionData: PositionData;
+}
+
+export interface UnrealizedRevenue {
+  unrealizedRevenue24hChangePercent: string;
+}
+
+export interface LoginAccount {
+  address: string;
+  displayAddress: string;
+  meta: { accontType: string; address: string; signer: object | null };
+  totalFrozenFunds: string;
+  tradingPositionsTotal: UnrealizedRevenue;
+  eth: string;
+  rep: string;
+  dai: string;
+}
+
+export interface BaseAction {
+  type: string;
+  data: any | undefined;
+}
